Remove NumberValidator dependency on brick/math
The min, max and step parameters are now provided as decimal strings, but internal calculations are performed on floats.
This seems to be good enough for now, as it looks like Chrome is doing the same in its current implementation or number inputs.

We may perform exact calculations again in the future, but the validator complexity will probably increase quite a lot.